scripts/requirements.txt
openai
anthropic
google-genai
python-pptx>=0.6.21
python-dotenv
pytest
